refactor: moving header to a folder
