+++ packages/react-router-v7-example/app/routes/index/route.tsx
@@ -1,77 +1,15 @@
 import type { Route } from './+types/route';
 
-import { Braces, Code, Eye, PanelLeft, PanelRight, SquarePen } from 'lucide-react';
-import { useEffect, useRef, useState } from 'react';
 import { InspectorSidebar } from '~/components/inspector-sidebar';
 import { TemplateSidebar } from '~/components/template-sidebar';
-import { Button } from '~/components/ui/button';
-import {
-  setSelectedMainTab,
-  toggleInspectorSidebarOpen,
-  toggleTemplatesSidebarOpen,
-  type MainTabOptions,
-} from '~/context/editor';
 import { Canvas } from './canvas';
-import { ExtraFunctions } from './extraFunctions';
-import { ScreenToggle } from './screenToggle';
+import { Header } from './header';
 
 export function meta({}: Route.MetaArgs) {
   return [{ title: 'New React Router App' }, { name: 'description', content: 'Welcome to React Router!' }];
 }
 
-const headerOptions: { label: MainTabOptions; icon: React.ElementType }[] = [
-  { label: 'editor', icon: SquarePen },
-  { label: 'preview', icon: Eye },
-  { label: 'html', icon: Code },
-  { label: 'json', icon: Braces },
-];
-
 export default function Home() {
-  const [hoveredIndex, setHoveredIndex] = useState<number | null>(null);
-  const [activeIndex, setActiveIndex] = useState(0);
-  const [hoverStyle, setHoverStyle] = useState({});
-  const [activeStyle, setActiveStyle] = useState({ left: '0px', width: '0px' });
-  const tabRefs = useRef<(HTMLDivElement | null)[]>([]);
-
-  const currentOption = headerOptions[activeIndex || 0];
-
-  useEffect(() => {
-    if (hoveredIndex !== null) {
-      const hoveredElement = tabRefs.current[hoveredIndex];
-      if (hoveredElement) {
-        const { offsetLeft, offsetWidth } = hoveredElement;
-        setHoverStyle({
-          left: `${offsetLeft}px`,
-          width: `${offsetWidth}px`,
-        });
-      }
-    }
-  }, [hoveredIndex]);
-
-  useEffect(() => {
-    const activeElement = tabRefs.current[activeIndex];
-    if (activeElement) {
-      const { offsetLeft, offsetWidth } = activeElement;
-      setActiveStyle({
-        left: `${offsetLeft}px`,
-        width: `${offsetWidth}px`,
-      });
-    }
-  }, [activeIndex]);
-
-  useEffect(() => {
-    requestAnimationFrame(() => {
-      const firstElement = tabRefs.current[0];
-      if (firstElement) {
-        const { offsetLeft, offsetWidth } = firstElement;
-        setActiveStyle({
-          left: `${offsetLeft}px`,
-          width: `${offsetWidth}px`,
-        });
-      }
-    });
-  }, []);
-
   return (
     <div className="flex min-h-screen max-h-svh overflow-hidden">
       {/* Left Sidebar */}
@@ -80,82 +18,11 @@ export default function Home() {
       {/* Main Content */}
       <div className="flex flex-col flex-1 min-w-0">
         {/* Header */}
-        <header className="flex h-14 items-center gap-2 border-b bg-background px-3 overflow-scroll">
-          <div className="flex items-center gap-3">
-            <Button variant="ghost" size="icon" className="h-8 w-8" onClick={toggleTemplatesSidebarOpen}>
-              <PanelLeft className="h-4 w-4" />
-            </Button>
-
-            {/* Dynamic Tabs */}
-            <div className="relative">
-              {/* Hover Highlight */}
-              <div
-                className="absolute h-[32px] transition-all duration-300 ease-out bg-muted/50 rounded-md flex items-center"
-                style={{
-                  ...hoverStyle,
-                  opacity: hoveredIndex !== null ? 1 : 0,
-                }}
-              />
-
-              {/* Active Indicator */}
-              <div
-                className="absolute bottom-[-6px] h-[2px] bg-foreground transition-all duration-300 ease-out"
-                style={activeStyle}
-              />
-
-              {/* Tabs */}
-              <div className="relative flex space-x-1 items-center">
-                {headerOptions.map((tab, index) => {
-                  const Icon = tab.icon;
-                  return (
-                    <div
-                      key={index}
-                      ref={(el) => {
-                        tabRefs.current[index] = el;
-                      }}
-                      className={`px-3 py-2 cursor-pointer transition-colors duration-300 h-[32px] flex items-center gap-2 ${
-                        index === activeIndex ? 'text-foreground' : 'text-muted-foreground'
-                      }`}
-                      onMouseEnter={() => setHoveredIndex(index)}
-                      onMouseLeave={() => setHoveredIndex(null)}
-                      onClick={() => {
-                        setActiveIndex(index);
-                        setSelectedMainTab(tab.label);
-                      }}
-                    >
-                      <Icon className="h-4 w-4" />
-                    </div>
-                  );
-                })}
-              </div>
-            </div>
-          </div>
-
-          <div className="flex-1" />
-
-          <div className="flex items-center gap-2">
-            {/* Screen Toggle Group */}
-            <ScreenToggle />
-
-            {/* Dropdown Menu */}
-            <ExtraFunctions />
-
-            {/* Save Button */}
-            <Button variant="ghost" size="sm">
-              SAVE
-            </Button>
-
-            <div className="h-4 w-px bg-border" />
-
-            <Button variant="ghost" size="icon" className="h-8 w-8" onClick={toggleInspectorSidebarOpen}>
-              <PanelRight className="h-4 w-4" />
-            </Button>
-          </div>
-        </header>
+        <Header />
 
         {/* Canvas Area */}
         <div className="flex-1 overflow-auto">
-          <Canvas canva={currentOption.label} />
+          <Canvas />
         </div>
       </div>
